Move reusable form components out of claim review folder
