+++ src/components/ClaimReview/form/fieldLists/reportedForm.ts
@@ -1,4 +1,4 @@
-import { createFormField, FormField } from "./FormField";
+import { createFormField, FormField } from "../../../Form/FormField";
 import { fetchUserList } from "./unassignedForm";
 
 const reportedForm: FormField[] = [
