+++ src/components/ClaimReview/form/fieldLists/unassignedForm.ts
@@ -1,6 +1,6 @@
 import usersApi from "../../../../api/userApi";
 import { Roles } from "../../../../types/enums";
-import { createFormField, FormField } from "./FormField";
+import { createFormField, FormField } from "../../../Form/FormField";
 
 export const fetchUserList = async (name, t) => {
     const userSearchResults = await usersApi.getUsers(
