+++ src/components/Form/SelectOptions.tsx
@@ -2,8 +2,8 @@ import { Select } from "antd";
 import { useTranslation } from "next-i18next";
 import React, { useMemo, useState } from "react";
 import styled from "styled-components";
-import colors from "../../../styles/colors";
-import Loading from "../../Loading";
+import colors from "../../styles/colors";
+import Loading from "../Loading";
 
 const StyledSelect = styled(Select)`
     .ant-select-selector {
